fix bugs discovered while testing on synthetix-v3
also includes some cleanliness refactors/organization.
