+++ packages/builder/src/util.ts
@@ -161,75 +161,3 @@ ${def.allActionNames.join('\n')}`);
 
   return str;
 }
-
-class ExtendableProxy {
-  readonly accessed = new Map<string, AccessRecorder>();
-
-  constructor() {
-    return new Proxy(this, {
-      get: (obj: any, prop: string) => {
-        // console.log('LOGGING ACCESS', typeof prop, prop);
-        if (prop === 'accessed' || prop === 'getAccesses') {
-          return obj[prop];
-        }
-        if (!this.accessed.has(prop)) {
-          this.accessed.set(prop, new AccessRecorder());
-        }
-        return typeof prop === 'symbol' ? () => '' : this.accessed.get(prop);
-      },
-    });
-  }
-}
-export class AccessRecorder extends ExtendableProxy {
-  getAccesses(depth: number, cur = 1) {
-    if (cur == depth) {
-      return Array.from(this.accessed.keys());
-    }
-
-    const acc: string[] = [];
-    for (const k in this.accessed.keys()) {
-      acc.push(k);
-      acc.push(
-        ...this.accessed
-          .get(k)!
-          .getAccesses(depth, (cur || 1) + 1)
-          .map((a) => `${k}.${a}`)
-      );
-    }
-
-    return acc;
-  }
-}
-
-export function computeTemplateAccesses(str?: string) {
-  if (!str) {
-    return [];
-  }
-  //const recorder = new AccessRecorder();
-  const recorders: { [k: string]: AccessRecorder } = {
-    contracts: new AccessRecorder(),
-    imports: new AccessRecorder(),
-    extras: new AccessRecorder(),
-    txns: new AccessRecorder(),
-    settings: new AccessRecorder(),
-  };
-
-  const baseTemplate = _.template(str, {
-    imports: {
-      ...recorders,
-      ...ethers.constants,
-      ...ethers.utils,
-    },
-  });
-
-  baseTemplate();
-
-  const accesses: string[] = [];
-
-  for (const recorder of Object.keys(recorders)) {
-    accesses.push(...Array.from(recorders[recorder].accessed.keys()).map((a: string) => `${recorder}.${a}`));
-  }
-
-  return accesses;
-  //return recorder.getAccesses(2).filter((a) => !a.includes('.'));
-}
